moving views to app
