+++ x-pack/plugins/profiling/public/components/app/stack_traces_view/get_stack_traces_tabs.ts
@@ -8,9 +8,9 @@
 import { EuiPageHeaderContentProps } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { TypeOf } from '@kbn/typed-react-router-config';
-import { TopNType } from '../../../common/stack_traces';
-import { StatefulProfilingRouter } from '../../hooks/use_profiling_router';
-import { ProfilingRoutes } from '../../routing';
+import { TopNType } from '../../../../common/stack_traces';
+import { StatefulProfilingRouter } from '../../../hooks/use_profiling_router';
+import { ProfilingRoutes } from '../../../routing';
 
 export function getStackTracesTabs({
   path,
