+++ x-pack/plugins/profiling/public/components/app/stack_traces_view/index.tsx
@@ -7,18 +7,18 @@
 import { EuiButton, EuiButtonGroup, EuiFlexGroup, EuiFlexItem, EuiPanel } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import React from 'react';
-import { StackTracesDisplayOption, TopNType } from '../../../common/stack_traces';
-import { groupSamplesByCategory, TopNResponse } from '../../../common/topn';
-import { useProfilingParams } from '../../hooks/use_profiling_params';
-import { useProfilingRouter } from '../../hooks/use_profiling_router';
-import { useProfilingRoutePath } from '../../hooks/use_profiling_route_path';
-import { useTimeRange } from '../../hooks/use_time_range';
-import { useTimeRangeAsync } from '../../hooks/use_time_range_async';
-import { AsyncComponent } from '../async_component';
-import { ChartGrid } from '../chart_grid';
-import { useProfilingDependencies } from '../contexts/profiling_dependencies/use_profiling_dependencies';
-import { ProfilingAppPageTemplate } from '../profiling_app_page_template';
-import { StackedBarChart } from '../stacked_bar_chart';
+import { StackTracesDisplayOption, TopNType } from '../../../../common/stack_traces';
+import { groupSamplesByCategory, TopNResponse } from '../../../../common/topn';
+import { useProfilingParams } from '../../../hooks/use_profiling_params';
+import { useProfilingRouter } from '../../../hooks/use_profiling_router';
+import { useProfilingRoutePath } from '../../../hooks/use_profiling_route_path';
+import { useTimeRange } from '../../../hooks/use_time_range';
+import { useTimeRangeAsync } from '../../../hooks/use_time_range_async';
+import { AsyncComponent } from '../../async_component';
+import { ChartGrid } from '../../chart_grid';
+import { useProfilingDependencies } from '../../contexts/profiling_dependencies/use_profiling_dependencies';
+import { ProfilingAppPageTemplate } from '../../profiling_app_page_template';
+import { StackedBarChart } from '../../stacked_bar_chart';
 import { getStackTracesTabs } from './get_stack_traces_tabs';
 
 export function StackTracesView() {
